farmpulse/backend/requirements.txt
fastapi
uvicorn
earthengine-api
transformers
Pillow
torch
# xgboost
# scikit-learn
# pandas
